Local changes to build latest weston for Ubuntu Xenial
Logged notes in build_notes, had to play with meson_options.txt
and basically started disabling instead of adding on components.
The only one I kept was gbm for renderer_gl, but I did not upgrade
from 18.<smthing> for xenial to 21.<smthing>, and in the process
I downgraded simple_clients to not use dmabuf-feedback, etc.

Installs all weston executable to my $WLD but does not launch yet..